funciona el server y el cliente

+++ socket.h
@@ -21,7 +21,7 @@ int socket_shutdown(socket_t* listener,int modo);
 int socket_bind_and_listen(socket_t* self, const char* host, const char* service);
 int socket_accept(socket_t* listener, socket_t* peer);
 int socket_connect(socket_t *self, const char *host, const char *service);
-ssize_t socket_send(socket_t* self, const char* buffer, size_t length);
+int socket_send(socket_t* self, const char* buffer, size_t length);
 ssize_t socket_receive(socket_t* self, char* buffer, size_t length);
 
 #endif
